Wrap register page search params in Suspense
