+++ app/api/shares/route.ts
@@ -12,7 +12,6 @@ export async function POST(req: NextRequest) {
   if (error || !data?.slug) {
     return NextResponse.json({ error: error?.message || 'slug missing' }, { status: 500 })
   }
-  const origin = process.env.NEXT_PUBLIC_SITE_URL || new URL(req.url).origin
   const origin = process.env.NEXT_PUBLIC_SITE_URL || req.headers.get('origin') || ''
   const url = `${origin}/s/${data.slug}`
   return NextResponse.json({ slug: data.slug, url })
